Create carpet components and fix the rutes
