+++ frontend/src/App.js
@@ -1,12 +1,12 @@
 
 import { BrowserRouter as Router, Routes, Route} from "react-router-dom";
-import { Login } from "./login/login";
-import { SelectRegister } from "./selectRegister/selectRegister";
-import{ StudentRegister } from "./studentRegister/studentRegister";
-import{ RenterRegister } from "./renterRegister/renterRegister";
-import { MainScreen } from "./inside/mainScreen";
-import{RenterUpdate} from "./update/renterUpdate";
-import{ StudentUpdate } from "./update/studentUpdate";
+import { Login } from "./components/login/login";
+import { SelectRegister } from "./components/selectRegister/selectRegister";
+import{ StudentRegister } from "./components/studentRegister/studentRegister";
+import{ RenterRegister } from "./components/renterRegister/renterRegister";
+import { MainScreen } from "./components/inside/mainScreen";
+import{RenterUpdate} from "./components/update/renterUpdate";
+import{ StudentUpdate } from "./components/update/studentUpdate";
 export function App() {
   return (
