+++ frontend/src/components/login/card.jsx
@@ -1,7 +1,7 @@
 import { CardBody } from "./cardBody";
 import { CardFooter } from "./cardFooter";
 import { CardHeader } from "./cardHeader";
-import styles from "../css/card.module.css"
+import styles from "../../css/card.module.css"
 
 export function Card(){
     return(
