+++ frontend/src/components/renterRegister/linksPage.jsx
