+++ frontend/src/components/renterRegister/renterRegister.jsx
@@ -1,12 +1,13 @@
 import { Link } from "react-router-dom";
-import styles from "../css/renterRegister.module.css";
-import { ButtonRegister } from "../renterRegister/buttonRegister";
-import { Description } from "../renterRegister/description";
-import { FormGroup } from "../renterRegister/formGroup";
-import { LinksPage } from "../renterRegister/linksPage";
-import { TitleRegister} from "../renterRegister/titleRenterRegister";
+import styles from "../../css/renterRegister.module.css";
+import { ButtonRegister } from "./buttonRegister";
+import { Description } from "./description";
+import { FormGroup } from "./formGroup";
+import { LinksPage } from "./linksPage";
+import { ListGenre } from "./selectGenre";
+import { TitleRegister} from "./titleRenterRegister";
 
-export function RenterUpdate() {
+export function RenterRegister() {
   return (
     <>
 
@@ -18,9 +19,9 @@ export function RenterUpdate() {
 
         <div className="row main">
 
-          <TitleRegister title="Renter Update"/>
+          <TitleRegister title="Renter Register"/>
 
-          <div className={styles.mainLoginUpdate}>
+          <div className={styles.mainLogin}>
             <form className="form-horizontal" method="post" action="#">
 
               <div className={styles.formGroup}>
@@ -38,13 +39,21 @@ export function RenterUpdate() {
               <div class={styles.formGroup}>
                 <Description/>
               </div>
+              
+              <div className={styles.formGroup}>
+                <FormGroup name="username" placeholder="Enter your Username" title="Username" image="fa fa-users fa" type="text"/>
+              </div>
+
+              <div className={styles.formGroup}>
+                <ListGenre/>
+              </div>
 
               <div className={styles.formGroup}>
                 <FormGroup name="confirm" placeholder="Enter your password" title="Password" image="fa fa-lock fa" type="password"/>
               </div>
 
               <div className={styles.formGroup}>
-                <ButtonRegister loginButton="Actualizar"/>
+                <ButtonRegister loginButton="Register"/>
               </div>
               
               <div className={styles.loginRegister}>
