+++ frontend/src/components/update/renterUpdate.jsx
@@ -1,13 +1,12 @@
 import { Link } from "react-router-dom";
-import styles from "../css/renterRegister.module.css";
-import { ButtonRegister } from "./buttonRegister";
-import { Description } from "./description";
-import { FormGroup } from "./formGroup";
-import { LinksPage } from "./linksPage";
-import { ListGenre } from "./selectGenre";
-import { TitleRegister} from "./titleRenterRegister";
+import styles from "../../css/renterRegister.module.css";
+import { ButtonRegister } from "../renterRegister/buttonRegister";
+import { Description } from "../renterRegister/description";
+import { FormGroup } from "../renterRegister/formGroup";
+import { LinksPage } from "../renterRegister/linksPage";
+import { TitleRegister} from "../renterRegister/titleRenterRegister";
 
-export function RenterRegister() {
+export function RenterUpdate() {
   return (
     <>
 
@@ -19,7 +18,7 @@ export function RenterRegister() {
 
         <div className="row main">
 
-          <TitleRegister title="Renter Register"/>
+          <TitleRegister title="Renter Update"/>
 
           <div className={styles.mainLogin}>
             <form className="form-horizontal" method="post" action="#">
@@ -39,21 +38,13 @@ export function RenterRegister() {
               <div class={styles.formGroup}>
                 <Description/>
               </div>
-              
-              <div className={styles.formGroup}>
-                <FormGroup name="username" placeholder="Enter your Username" title="Username" image="fa fa-users fa" type="text"/>
-              </div>
-
-              <div className={styles.formGroup}>
-                <ListGenre/>
-              </div>
 
               <div className={styles.formGroup}>
                 <FormGroup name="confirm" placeholder="Enter your password" title="Password" image="fa fa-lock fa" type="password"/>
               </div>
 
               <div className={styles.formGroup}>
-                <ButtonRegister loginButton="Register"/>
+                <ButtonRegister loginButton="Actualizar"/>
               </div>
               
               <div className={styles.loginRegister}>
