+++ frontend/src/css/titleStudentRegister.module.css
@@ -1,5 +1,5 @@
 h1.title {
-    transform: translate(115%, 25%);
+    transform: translate(120%, 25%);
     font-size: 50px;
     font-family: "Passion One", cursive;
     font-weight: 400;
